Add script to send messages from the CLI to Slack
Used this to bulk send messages to 30 people. It was slow but I did get
individual messages out and it was faster than doing it manually.

The script on the CLI had a 2s sleep and I had to modify the original
slack_applescript to send newlines as separate outputs because
otherwise it didn't send the full message.