Allow for overrides on which commit for PRs
I kept going in and changing the title on the PR, so let's just make
that a part of the flow if we have more than 1 commit, so I can decide
which one is the best one.

Because often I do something that is just two commits, and the first
one perfectly describes the PR, and the second is some related but not
primary task I did. Or refactor.